Continued swatting at bugs with too small a flyswatter while continuing to migrate to Codable from what I had before. (Which worked, but sucked in its way.)
* Started migrating to John Sundell's Codextended as that offers a nicer front end.
* Renamed PersistenceController to just Persistence because I'm daring that way.

I now think I know how I'm going to finish the whole Codable implementation problem: I *think* (but need to check the postgresql implementation) that the wrapper for each record in the bestiary import is not a necessary detail. So not having to store it would simplify a lot: look up the model name, hand the content of fields off to the right class.